requirements.txt
streamlit
pandas
google-cloud-translate
openpyxl
tempfile
os
